* Changed constant length for MUSL

+++ src/constants.rs
@@ -12,7 +12,7 @@ pub type ConstantType = u64;
 pub type ConstantType = i32;
 
 #[cfg(all(target_pointer_width = "64", target_env = "musl"))]
-pub type ConstantType = i32;
+pub type ConstantType = u64;
  
 #[cfg(all(target_pointer_width = "32", not(target_env = "musl")))]
 pub type ConstantType = u32;
